requirement.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.2
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
resend==2.7.0
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0